fix: inference pipeline correctness issues
Problem 1 - numpy/torch type error:
dt_dataset.py raised TypeError when assigning a numpy.float32 RTG
value into a torch.FloatTensor slice. Affected all training runs
that included the new Game 1-9 episodes.

Problem 2 - out-of-hand card recommendations:
DTStrategy.recommend() built affordable_indices from the raw
hand_cards list which contains towers (princess-tower, king-tower)
and -next preview slots alongside actual -in-hand cards. Towers
have elixir cost 0 so were already excluded, but -next cards have
real costs and were leaking through, causing the strategy to
sometimes recommend a card not yet in the player's hand.

Problem 3 - --no-cache deleted analysis files:
run_inference.py had a --no-cache flag that deleted the cached
analysis JSON before re-running. This destroyed patched analysis
files irreversibly. The flag served no legitimate use case.

What we did:
- Cast RTG value to float() before torch tensor assignment.
- Added "-in-hand" in card guard to affordable_indices filter so
  only cards the player can actually play are considered.
- Removed --no-cache and --cached-analysis flags from
  run_inference.py entirely.

Known remaining issue:
Roboflow frequently detects fewer than 4 -in-hand cards per frame
(only 10% of frames have all 4 detected). The hand_cards field
in analysis JSONs is therefore sparse and unreliable. A HandTracker
will be implemented in a separate branch to fix this by propagating
hand state across frames using play-event detection.

+++ src/data/dt_dataset.py
@@ -108,7 +108,7 @@ class DecisionTransformerDataset(Dataset):
 
             # Normalized RTG
             raw_rtg = ep.returns_to_go[start + i]
-            returns_to_go[pad + i, 0] = (raw_rtg - self.rtg_mean) / self.rtg_std
+            returns_to_go[pad + i, 0] = float((raw_rtg - self.rtg_mean) / self.rtg_std)
 
         # Data augmentation (training only, controlled by std > 0)
         if self.state_noise_std > 0:
